sanitized parameters for devise and removed registration controller
