+++ wardesportsapp/Gemfile
@@ -32,7 +32,11 @@ gem 'jbuilder', '~> 2.5'
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
 
+# Sign up & Login Engine
 gem 'devise'
+
+# Debugging
+gem 'pry-rails'
 
 # Use Capistrano for deployment
 # gem 'capistrano-rails', group: :development
