--- wardesportsapp/app/helpers/registrations_helper.rb
@@ -1,2 +1,0 @@
-module RegistrationsHelper
-end
